pyproject.toml
[tool.poetry]
name = "lastfm-count"
version = "0.1.0"
description = ""
authors = ["Josh Finnie <josh@jfin.us>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.27.1"
python-dotenv = "^0.19.2"
twitter = "^1.19.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
